goog_trends/requirements.txt
requests
beautifulsoup4
pandas
yfinance
matplotlib
seaborn
numpy
python-dotenv
pytrends
